Version 1.2.5
Cambios en las variantes byte de telefonoalumno y de fechadeNacimientoAlumno a localDate y int

+++ src/servicios/AlumnoImplementacion.java
@@ -34,7 +34,7 @@ public class AlumnoImplementacion implements AlumnoInterfaz {
 		String gmailAlumno=scanner.next();
 
 		System.out.println("Telefono");
-		byte telefonoAlumno=scanner.nextByte();
+		int telefonoAlumno=scanner.nextInt();
 		
 		
 	}
